navigation: adds convertion function from return code to string
