+++ px4_ros2_cpp/include/px4_ros2/navigation/experimental/global_navigation_interface.hpp
@@ -11,7 +11,6 @@
 
 #include <px4_msgs/msg/aux_global_position.hpp>
 #include <px4_ros2/navigation/experimental/navigation_interface_base.hpp>
-#include <px4_ros2/navigation/experimental/navigation_interface_common.hpp>
 
 using namespace Eigen;
 using namespace px4_msgs::msg;
